feat: Slack backend plugin (API-4289) (#2)
* Add a non-functional backend slack API

This is the initial commit for a backend Slack API for our plugin
building exploration.

Right now the slack API only contains two endpoints
/api/slack/health and /api/slack/chat.

Neither endpoint does much. The health endpoint returns a status of 'ok'
and logs to the server log. The slack router does the same.

I haven't particularly followed any patterns with this plugin
development. By nature I think these plugins are going to be small-ish.
I think as we move along we may want to consider following common
setups. (Routes folder versus a controller folder - where business logic
should reside, etc.)

API-4289

* Read slack token from environment variables

The slack plugin can now pull the slack token from the SLACK_TOKEN
environment variable.

In doing this I added an environment helper, that may or may not be
useful in the future, and modified the controllers to take the
'PluginEnvironment' type rather than the RouterOptions type I created in
an earlier commit - I removed that RouterOptions type since it was
unused after this change.

API-4289

* Remove PluginEnvironment import

PluginEnvironment is defined within backstage, which doesn't give us
nice access to it in plugins - technically we can get since we're in a
monorepo, but if we were to move plugins that would caused issues. To
fix this I created a plugin specific interface for the Slack environment
options, similar to the one I removed in the last commit.

This new interface grabs the Config and Logger types directly from their
packages rather than depending on the arbitrary type defined in the
backend folder.

Note: There's an issue with my linter in the backend folder complaining
that @internal/plugin-slack-backend' is not declared in my dependencies.
Since we don't have the slack plugin published I can't resolve this. The
code works, again, because we have the monorepo, but this lint error is
still there.

API-4289

* Add some TODOs

* Fix main slack router unit test

* Add e2e test for slack plugin

API-4289

* Add unit tests for controllers

I added unit tests for controllers. Since there's no deeper
implementation right now they look like excerpts from the e2e test. This
will change as we add in the actual slack functionality though.

API-4289

* Add comments explaining the need to mock the config

* Resolve error in standalone server on slack plugin

I wasn't able to test this because I think standalone server might only
be for kubernetes instances. I'm not certain. This change doesn't cause
errors for the regularly running app and the tests still pass. However,
this does fix the lint error that createRouter requires a config object.

API-4289

* Fix bad import

I had an import from the backend, but I shouuld have been importing from
@backstage.

API-4289

* Slack now posts successfully based on environment

Slack is now posting successfully based on the given configuration for
the bot id, base url, token, and channel

API-4289

* Send message received instead of a static one

The /postMessage endpoint now takes the message to be sent rather than
sending a static one each time.

API-4289

* Add basic slack client test

API-4289

* Update slack plugin README

Added configuration options and endpoints to the slack plugin README

API-4289

* Update tests to match latest code

I didn't have tests updated for the latest code changes

API-4289

* Rename e2e test to int and resolve errors

The slack plugin router e2e test is now called 'int' for integration,
since it isn't really reaching out to Slack.

I also received issues with it not passing.

API-4289

* Clean up test codes and comments

API-4289

* Add example frontend slack plugin

Added a frontend slack plugin for demoing slack. Note that this is only
for demoing.

API-4289

* Remove internal import

Internal imports are not appreciated in this repo so I took one out and
changed the type to Record<string, unknown> for now.

API-4289

* Provide mock discovery api service for plugin test

The SlackAPI requires the discoveryApi. This adds the needed
discoveryApi via the backstage utility functions.

Note: I'm encountering an error that canvas.getContext() is not defined
with jsdom. I'm searching for a fix but am curious if this a recent
backstage change/package update causing this so left it for now.

API-4289

* Add clarifying comment

* Add slack plugin configuration instructions

* Resolve some tsc unused value errors

There were a couple instances that tsc was throwing errors in circleCI
because some arguments in function signatures were not being used

API-4289

* Add slack options to .env.sample

* Add missing internal dependencies to packagejsons

The internal dependencies (@internal/package-name) were not included in
the package.jsons. While the app compiles successfully and runs, we
should treat them as if they were from separate repositories and still
list those dependencies in our package.json.

API-4289

* Complete an unfinished comment

* Add README file for creating custom plugins

WIth the current backstage.io not including a page specifically for
backend plugins (says TODO at the moment), I added a README to describe
how we can create backend plugins and integrate a frontend plugin with
it.

* Add note about using the discovery api

* Add link to base README for backend plugin guide

* Set docker image node to 14.15.4

This version of node is required for backstage to run at this time, so
we need to hardset our docker container to that version.

API-4289